#35 Refactor query and mutation
