+++ Pronto-MIA/Pronto-MIA/Services/GraphQLService.cs
@@ -13,6 +13,7 @@ namespace Pronto_MIA.Services
     using Pronto_MIA.BusinessLogic.API.Logging;
     using Pronto_MIA.BusinessLogic.API.Types;
     using Pronto_MIA.BusinessLogic.API.Types.Mutation;
+    using Pronto_MIA.BusinessLogic.API.Types.Query;
 
     /// <summary>
     /// Service which initializes and contains all information regarding the
@@ -37,7 +38,12 @@ namespace Pronto_MIA.Services
                 .AddType<DeploymentPlanResolvers>()
                 .AddType<AccessControlListInputType>()
                 .AddQueryType<Query>()
+                .AddTypeExtension<DeploymentPlanQuery>()
+                .AddTypeExtension<UserQuery>()
+                .AddTypeExtension<DepartmentQuery>()
                 .AddMutationType<Mutation>()
+                .AddTypeExtension<DeploymentPlanMutation>()
+                .AddTypeExtension<FcmTokenMutation>()
                 .AddTypeExtension<UserMutation>()
                 .AddTypeExtension<DepartmentMutation>()
                 .AddProjections()
